Rebase and improve trace overview output
This commit is a rebase that adds minor tweaks to allow the
overview output to be part of the persisting state of the
trace context and improves the positioning of the overview
output while using in conjunction a pinned view. It also fixes
a UI bug related to closing the overview output, and reorders
the trace-context-component toolbar.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>